feat(dialog): add post-game dialog message
The dialog notifies you of a victory, defeat, draw, or forfeiture. This needs
a tiny bit of extra work, because the clients are still unaware of each other
in terms of profile details, even though the server handles this. So, the
next step is to make each dialog mention the appropriate player's name.